requirements.txt
# LLM
openai>=1.0.0
anthropic>=0.18.0

# RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
chromadb>=0.4.0

# Document processing
pypdf>=3.0.0
arxiv>=2.0.0

# Utils
python-dotenv>=1.0.0

# UI 
streamlit>=1.30.0
